requirements.txt
flask
redis
python-consul
